last batch

+++ vm/message/src/lib.rs
@@ -31,7 +31,7 @@ pub trait Message {
     /// sets a new sequence to the message.
     fn set_sequence(&mut self, sequence: u64);
     /// Returns the gas limit for the message.
-    fn gas_limit(&self) -> i64;
+    fn gas_limit(&self) -> u64;
     /// Returns the required funds for the message.
     fn required_funds(&self) -> TokenAmount;
     /// gets gas fee cap for the message.
@@ -63,8 +63,8 @@ impl Message for ShimMessage {
     fn params(&self) -> &RawBytes {
         &self.params
     }
-    fn gas_limit(&self) -> i64 {
-        self.gas_limit as i64
+    fn gas_limit(&self) -> u64 {
+        self.gas_limit
     }
     fn set_gas_limit(&mut self, token_amount: i64) {
         self.gas_limit = token_amount as u64;
